libretro: Gracefully shutdown when closing content.
The core had a tendency of closing even the frontend itself due to
the exit() command. It should now only close itself when the user
asks the frontend to close the running content.

However, this exposed another problem. If the user requests a
content restart, the frontend freezes. Before, it would basically kill
itself because of the previous exit behaviour, but now it freezes
indefinitely. This needs to be looked at later... in the worst
case, the restart function will need to be stubbed.