freej2me-lr: Fix rotated pointer coordinates.
Before, when the screen was rotated on the libretro frontend, the
pointer coordinates in the java file were reversed correctly, but
some extra calculations to account for the 90 degrees
counterclockwise rotation were not done, making the pointer be
"reversed" in the y axis, so the lower the pointer was in the
screen, the higher the y value it had in the rotated mode. Now it
works as intended.