mobile: PlatformPlayer: Add a catch for NULL wavClips.
Some games loadd a NULL pointer as a wavClip, then try to check if
that wavClip is running, which in turn makes FreeJ2ME throw an
exception. This workaround should help games in those situations,
and also refrain from causing problems on others by messing with
the media's getState() as a whole.