mobile: PlatformPlayer: Re-introduce audio/wav with a try-catch.
We're going to use audio/wav type, even if some of those are
encoded in the ADPCM format. For cases where an unsupported format
is loaded, the improved try-catch statement on the wavPlayer's
constructor will now warn about the problem, and close the wavClip
to prevent it from being used. Some games like Motocross Trial
Extreme, however, keep trying to load those wav files
indefinitely even with the try-catch closing them, but it should
be fixed when the playback of ADPCM wav gets implemented.