mobile: PlatformPlayer: Smaller improvements to the midiPlayer.
Certain games like Star Warriors (fan-translation of a chinese
port of Raiden to J2ME phones) keep hammering midiPlayer's start()
method all the time during gameplay, resulting in a constant
repeat of the first second or so of the midi stream. Adding a
check for isRunning() in there like it's done for wavPlayer solves
this issue, and makes sense because to play tha stream again, it
should be stopped or paused first.

A change that allows midiPlayer's getMediaTime() method to return
the midi's current time was also implemented, by returning
midi.getTickPosition().