libretro: Port to Win32 and introduce basic logging support.
That one took a while... but at least now it works, and it is also
on par with the linux core having all the newest improvements and
fixes. Tested only on Windows 10 x64.

To build for win32, you will need mingw or MSYS2 UCRT64 (which i
used). With UCRT64 it's a bit easier since you only need to cd
into src/libretro, run "make" and it will give you the core's dll.

This commit also introduces support for logging, allowing the core
to be debugged if needed, although most of the logs are centered
on win32.