ImaAdpcmDecoder: More adjustments to IMA ADPCM decoder
This is an additional pass to further cleanup and improve the code.

Checking if the fourth preamble byte is '!= 0' has no use here and
does more harm than good, and in cases the preamble is read
one or more times, the decoder accounts for it on the resulting
stream's size for better time and space efficiency by eliminating
the need to copy the data into a new correctly sized byte array.

IMA ADPCM's header is also fully stripped from the decoding
process as well, it is only used to get information regarding
format, number of channels, bit rate and so on.