ImaAdpcmDecoder: Even more refinements to the decoder.
Time for some final touches on this one. Header creation was
reworked, decoding algorithm is now much better with little to
no crackling even on small streams, everything that isn't going
to change (and arrays that aren't going to be reassigned) are now
final for better performance, and the code is a bit more readable
as well.